pyproject.toml
[project]
name = "face-fft"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "diffusers>=0.37.0",
    "imageio[ffmpeg]>=2.37.3",
    "pillow>=12.1.1",
    "scikit-learn>=1.8.0",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
    "transformers>=5.3.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
